S68 Wave 6: 25 more f32 shaders evolved to f64 canonical
Interpolation (1): loo_cv
Loss (20): center_loss, chamfer_distance, contrastive_loss,
cosine_embedding_loss, dice_loss, earth_mover_distance,
focal_loss, focal_loss_alpha, focal_loss_v2, giou_loss,
iou_loss, kldiv_loss, label_smoothing, lovasz_loss, nll_loss,
perceptual_loss, poisson_nll_loss, triplet_loss, tversky_loss,
wasserstein_loss
Math (4): acosh, acos, argmax_reduce, argmax

acosh: Implemented as log_f64(x + sqrt_f64(x*x - 1.0)) since no
native acosh_f64 polyfill exists.

700 WGSL shaders, 0 clippy warnings.

Made-with: Cursor

+++ crates/barracuda/src/shaders/interpolation/loo_cv_f64.wgsl
@@ -1,12 +1,12 @@
-// Leave-One-Out Cross-Validation for kernel methods
+// loo_cv_f64.wgsl — Leave-One-Out Cross-Validation for kernel methods (f64 canonical)
 // LOO_i = (y_i - pred_i) / (1 - H_ii)
 // H: hat matrix [N, N], y: targets [N], pred: predictions [N]
 // H stored row-major, H[i,i] = H[i*N + i]
 
-@group(0) @binding(0) var<storage, read> hat_matrix: array<f32>;
-@group(0) @binding(1) var<storage, read> y: array<f32>;
-@group(0) @binding(2) var<storage, read> predictions: array<f32>;
-@group(0) @binding(3) var<storage, read_write> output: array<f32>;
+@group(0) @binding(0) var<storage, read> hat_matrix: array<f64>;
+@group(0) @binding(1) var<storage, read> y: array<f64>;
+@group(0) @binding(2) var<storage, read> predictions: array<f64>;
+@group(0) @binding(3) var<storage, read_write> output: array<f64>;
 @group(0) @binding(4) var<uniform> params: Params;
 
 struct Params {
@@ -25,7 +25,7 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
     let denom = 1.0 - h_ii;
 
     // Avoid division by zero (H_ii = 1 means point has full influence - edge case)
-    var loo_residual: f32;
+    var loo_residual: f64;
     if (abs(denom) < 1e-10) {
         loo_residual = 0.0;
     } else {
