+++ crates/barracuda/src/shaders/loss/earth_mover_distance_f64.wgsl
@@ -1,4 +1,4 @@
-// earth_mover_distance.wgsl - Earth Mover's Distance (Wasserstein-1)
+// earth_mover_distance_f64.wgsl - Earth Mover's Distance (Wasserstein-1) (f64 canonical)
 //
 // Measures distance between probability distributions
 // Also known as Wasserstein distance
@@ -16,12 +16,12 @@ struct Params {
     _pad6: u32,
 }
 
-@group(0) @binding(0) var<storage, read> dist1: array<f32>;     // Distribution 1 (probabilities)
-@group(0) @binding(1) var<storage, read> dist2: array<f32>;     // Distribution 2 (probabilities)
-@group(0) @binding(2) var<storage, read_write> output: array<f32>; // Scalar distance
+@group(0) @binding(0) var<storage, read> dist1: array<f64>;     // Distribution 1 (probabilities)
+@group(0) @binding(1) var<storage, read> dist2: array<f64>;     // Distribution 2 (probabilities)
+@group(0) @binding(2) var<storage, read_write> output: array<f64>; // Scalar distance
 @group(0) @binding(3) var<uniform> params: Params;
 
-var<workgroup> shared_emd: array<f32, 256>;
+var<workgroup> shared_emd: array<f64, 256>;
 
 @compute @workgroup_size(256)
 fn main(
@@ -30,21 +30,21 @@ fn main(
 ) {
     let idx = global_id.x;
     let local_idx = local_id.x;
-    
+
     // Compute cumulative distributions (CDF)
     // For proper EMD, we need CDF differences
-    
-    var local_emd: f32 = 0.0;
-    
+
+    var local_emd: f64 = 0.0;
+
     if (idx < params.size) {
         // Simplified: sum absolute differences (approximation)
         // Full EMD requires sorting and transport plan
         local_emd = abs(dist1[idx] - dist2[idx]);
     }
-    
+
     shared_emd[local_idx] = local_emd;
     workgroupBarrier();
-    
+
     // Parallel reduction
     var stride = 128u;
     while (stride >= 1u) {
@@ -54,7 +54,7 @@ fn main(
         workgroupBarrier();
         stride = stride / 2u;
     }
-    
+
     if (local_idx == 0u) {
         output[0] = shared_emd[0];
     }
